Breaking rewrite in TypeScript
This change is a total rewrite of the `quill-cursors` module that
attempts to address a number of things in this one refactor.

This change makes some breaking changes:

  - `setCursor` method removed in favour of using `createCursor` and
    then `moveCursor` (see the example)
  - CSS is now in-lined in the JavaScript file, so there is no need to
    separately include styles
  - configuration of the module has changed (see the `README`)
  - the `quill-cursors` module will now - by default - emit a
    `selection-change` event whenever a `text-change` is fired

One of the bigger changes is that this module is now written in
TypeScript. This leads to a nicer development experience with some
compile-time checks.

The project will still transpile down to ES5.

The previous approach to updating cursors had two separate events it was
concerned with:

  - `selection-change`: when a user manually moves their cursor, then
    we need to update their cursor position with the selection we get
    from this event
  - `text-change`: Quill does not emit a `selection-change` event when
    a user is typing, instead emitting a `text-change` event. This
    module used to try to parse this text change, and turn that into an
    appropriate shift for the cursors

The need to handle two types of a event is slightly brittle, and parsing
deltas should not be a concern of this library.

We now move to a new approach. When a user types and triggers a
`text-change` event, we also emit a `selection-change` event with their
current selection. This way, clients using this library still only need
to handle a `selection-change` event, but they will also get more
accurate selections on `text-change`, using no delta parsing.

This change also makes a number of changes to the build:

  - 100% test coverage has been added with the Jest testing framework
  - Webpack has been bumped to v4 to address some security issues
  - Travis config has been added
  - `package-lock.json` has been removed, as clients will ignore this
    file when resolving their own dependency graphs