Add the cursor ID and a custom CSS class to the cursor container
Sometimes it can be useful to adjust the CSS class attached to the
cursor container. For example, when using multiple cursor modules on the
same document.

This change adds an option to add a custom CSS class to the container.
It also adds an `id` attribute to each container, in case you want to
target a particular cursor through either CSS or JavaScript.