Use `ResizeObserver` instead of listening for window `resize` event
We currently listen for `resize` events fired on the window to redraw
the cursors.

However, this approach misses the event where the editor is resized for
other reasons (such as through JavaScript), without the window being
resized.

Ideally, we'd use the new [`ResizeObserver`][1] class, but this is
restricted just to Chrome for now. Instead, we're using a [polyfill][2].

[1]: https://developers.google.com/web/updates/2016/10/resizeobserver
[2]: https://github.com/pelotoncycle/resize-observer