Add local cursor transforms on `text-change`
At the moment, we have to wait for clients to report their position
before we update it. This is fine on very low-latency connections (like
when developing on the same machine), but on high-latency connections,
it can lead to some laggy or juddery performance as the cursor drifts
out-of-sync with the `text-change` events received by Quill.

This change adds in a `text-change` listener that will transform all
cursors by the received delta using `quill-delta`'s `transformPosition`,
which is meant for exactly this sort of application.

The example has also been updated to show how a cursor implementation
might look on a high-latency connection, using a debounced update to
avoid laggy behaviour.