Use pointer-events: none for selections
We previously set the selections' `z-index: -1` to avoid the selections
blocking click events. However, this doesn't work on block-level
elements, such as `image-block`s, because the selection is perfectly
behind the image.

This change removes the `z-index` assignment, and we instead let the
selections remain on top, but we apply `pointer-events: none` to avoid
the selections reacting to `click` or `touch` events.