Build selection from multiple Ranges
We currently build a selection from a single `Range` object, which can
contain a mixture of `inline` and `block` elements. In particular, with
paragraphs, the behaviour between browsers is inconsistent.

For example, when selecting an entire paragraph, Chrome will draw
rectangles around each line in the paragraph, as well as a big rectangle
around the `<p>` element itself. Firefox, on the other hand, will only
draw a rectangle around the `<p>` block, but not the individual lines.
Both of these behaviours are different still to how the browser's
built-in "selection" looks (ie just drawing rectangles around the
individual lines, and never the `<p>` element).

In order to simulate this more "natural" selection behaviour, this
change leans a bit more on Quill, and requests all of the Quill `Block`s
in a given range. For each of these `Block`s, we create a DOM `Range`
object, which avoids ever selecting the parent `Block` (unless that
`Block` is an embed like an `image-block`, in which case it will be
selected).