Fix max Quill index
When we try to highlight a range, we first attempt to normalise it to
within Quill's bounds. However, the upper limit we apply is the length
of the Quill document, when actually - because Quill is zero-indexed -
we should set the upper limit to `quill.getLength() - 1`.

This change updates our upper limit so that we avoid getting errors when
attempting to set the selection outside the allowable range.