Hide flag immediately when actively toggling
This change updates our new flag toggling logic so that if a flag is
actively toggled, it is hidden immediately (ignoring the configured
`hideDelayMs` option). The hide delay should still apply in the hover
case.

We achieve this by adding a `.no-delay` class to the flag when we toggle
it, which suppresses the `transition-delay`. We have then wait for the
animation to complete before removing it again, so that the flag still
behaves as configured on hover.