Breaking: Remove core Quill styling
This change removes virtually all the styling we do on the "core" Quill
elements, in order to remain as unopinionated as possible.

We explicitly add `position: relative`, which technically isn't needed,
because it's included in Quill's default stylings. However, people are
free to ignore Quill's default stylings, or the styling could be removed
altogether, in which case our `absolute` positioning would break, so
let's explicitly set this just to be safe.

The one exception we have to make is that we need to set
`overflow: hidden` on `.ql-container`, which fixes the issue of
selections appearing outside an editor with a fixed height:
https://github.com/reedsy/quill-cursors/pull/21

Note that this should be acceptable, because even Quill's Bubble theme
displays poor behaviour in this case, so it can be considered a "fix".
Any consumers who care enough can override this, but it at least
provides us with sensible behaviour out-of-the-box.

This is technically a breaking change, in case anyone was relying on our
core styling.