Fix image selection
This change fixes a bug that can be reproduced by:

 1. Run the development app
 2. Paste an image into the editor
 3. Select all
 4. The "remote" cursor does not highlight the image, and the console
    errors

This is because when using `Range.setStart` and `Range.setEnd`, the
`offset` argument is treated differently for non-text nodes.

This change updates our range definitions to check if the leaf in
question is a text node or not. If it is a text node, then we use the
offset as before. However, if it isn't a text node (eg an `img`), then
we simply set the selection before/after the node using
`Range.setStartBefore` and `Range.setEndAfter`.