🏷️ Conform to Quill `Module` interface
Quill's [`Module` class][1] has a number of properties defined on it,
which the `QuillCursors` module doesn't expose.

This can make it tricky when trying to write generic type definitions
involving modules, since this module doesn't conform to the expected
interface.

This change aligns us with the `Module` interface. This also involves
including a static `DEFAULTS` property, which consumers could in theory
override themselves, so here we also populate our default options from
this property.

[1]: https://github.com/quilljs/quill/blob/develop/core/module.js